pyproject.toml
[project]
name = "pq-fido2-authenticator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cbor2>=5.6.0",
    "fido2>=2.0.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "keyring>=25.6.0",
    "liboqs-python>=0.14.1",
    "playwright>=1.55.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "pyobjc>=12.0; sys_platform == 'darwin'",
    "sqlalchemy>=2.0.44",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "ruff>=0.14.4",
]
